requirement.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-widget-tweaks==1.5.0
mysql-connector-python==9.6.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
folium==0.20.0
requests==2.32.3
gpxpy==1.6.2

pillow==10.1.0
